+++ lib/msf/ui/console/driver.rb
@@ -27,6 +27,13 @@ class Driver < Msf::Ui::Driver
   DefaultPromptChar = "%clr>"
 
   #
+  # Console Command Dispatchers to be loaded after the Core dispatcher.
+  #
+  CommandDispatchers = [
+    CommandDispatcher::Jobs
+  ]
+
+  #
   # The console driver processes various framework notified events.
   #
   include FrameworkEventManager
@@ -108,6 +115,10 @@ class Driver < Msf::Ui::Driver
       print_error("***")
     end
 
+    # Load the other "core" command dispatchers
+    CommandDispatchers.each do |dispatcher|
+      enstack_dispatcher(dispatcher)
+    end
 
     # Add the database dispatcher if it is usable
     if (framework.db.usable)
